requirements.txt
biopython==1.81
contourpy==1.1.0
cycler==0.11.0
fonttools==4.40.0
kiwisolver==1.4.4
matplotlib==3.7.1
nex==1.0.4
numpy==1.25.0
packaging==23.1
Pillow==10.0.0
pyparsing==3.1.0
python-dateutil==2.8.2
six==1.16.0
